requirements.txt
aiohttp==3.10.11
beautifulsoup4==4.12.3
anthropic==0.35.0
httpx==0.27.2
telethon==1.36.0
python-dotenv==1.0.1
curl_cffi==0.7.3
ddgs==9.12.0
pytz==2024.2
